Creating PushPublishigDependencyProcesor class
