+++ dotCMS/src/main/java/com/dotcms/publisher/util/dependencies/DependencyManager.java
@@ -87,13 +87,9 @@ import org.apache.commons.lang.StringUtils;
 public class DependencyManager {
 
 	private PublisherFilter publisherFilter;
-
-	private DependencyModDateUtil dependencyModDateUtil;
-	private PushedAssetUtil pushedAssetUtil;
 	private User user;
-
 	private PushPublisherConfig config;
-	private ConcurrentDependencyProcessor dependencyProcessor;
+	private PushedAssetUtil pushedAssetUtil;
 
 	@FunctionalInterface
 	public interface SupplierWithException<T> {
@@ -109,53 +105,8 @@ public class DependencyManager {
 	 */
 	public DependencyManager(User user, PushPublisherConfig config) {
 		this.config = config;
-		dependencyModDateUtil = new DependencyModDateUtil(this.config);
-		pushedAssetUtil = new PushedAssetUtil(config);
-
-		// these ones store the assets that will be sent in the bundle
-		dependencyModDateUtil = new DependencyModDateUtil(config);
-		dependencyProcessor = new ConcurrentDependencyProcessor();
-		this.config.setDependencyProcessor(dependencyProcessor);
 		this.user = user;
-
-		dependencyProcessor.addProcessor(PusheableAsset.SITE, (host) ->
-				proccessHostDependency((Host) host));
-
-		dependencyProcessor.addProcessor(PusheableAsset.FOLDER, (folder) ->
-				processFolderDependency((Folder) folder));
-
-		dependencyProcessor.addProcessor(PusheableAsset.TEMPLATE,
-				(template) -> processTemplateDependencies((Template) template));
-
-		dependencyProcessor.addProcessor(PusheableAsset.CONTAINER,
-				(container) -> processContainerDependency((Container) container));
-
-		dependencyProcessor.addProcessor(PusheableAsset.CONTENT_TYPE,
-				(contentType) -> processContentTypeDependency((Structure) contentType));
-
-		dependencyProcessor.addProcessor(PusheableAsset.LINK,
-				(link) -> processLinkDependency((Link) link));
-
-		dependencyProcessor.addProcessor(PusheableAsset.CONTENTLET, (content) -> {
-			try {
-				processContentDependency((Contentlet) content);
-			} catch (DotBundleException e) {
-				Logger.error(DependencyManager.class, e.getMessage());
-				throw new DotRuntimeException(e);
-			}
-		});
-		dependencyProcessor.addProcessor(PusheableAsset.RULE, (rule) -> setRuleDependencies((Rule) rule));
-		dependencyProcessor.addProcessor(PusheableAsset.LANGUAGE, (lang) -> {
-			try {
-				processLanguage((Language) lang);
-			} catch (DotBundleException e) {
-				Logger.error(DependencyManager.class, e.getMessage());
-				throw new DotRuntimeException(e);
-			}
-		});
-		dependencyProcessor.addProcessor(PusheableAsset.RELATIONSHIP,
-				(relationship) -> processRelationshipDependencies((Relationship) relationship));
-
+		pushedAssetUtil = new PushedAssetUtil(config);
 	}
 
 	/**
@@ -176,9 +127,8 @@ public class DependencyManager {
 		Logger.debug(DependencyManager.class,
 				"publisherFilter.isDependencies() " + publisherFilter.isDependencies());
 
-		if (publisherFilter.isDependencies()) {
-			setLanguageVariables();
-		}
+		this.config.setDependencyProcessor(
+				new PushPublishigDependencyProcesor(user, config, publisherFilter));
 
 		List<PublishQueueElement> assets = config.getAssets();
 
@@ -390,940 +340,10 @@ public class DependencyManager {
 	}
 
 	private <T> void addDirectly(final T asset, final PusheableAsset pusheableAsset) {
-		final boolean add = add(pusheableAsset, asset);
+		final boolean added = config.addWithDependencies(asset, pusheableAsset);
 
-		if (add) {
-			dependencyProcessor.addAsset(asset, pusheableAsset);
-		}
-	}
-
-	private Folder getFolderByParentIdentifier(final Identifier identifier)
-			throws DotDataException, DotSecurityException {
-
-		return APILocator.getFolderAPI()
-				.findFolderByPath(identifier.getParentPath(), identifier.getHostId(), user,
-						false);
-	}
-
-	private Host getHostById(final String hostId)
-			throws DotDataException, DotSecurityException {
-
-		return APILocator.getHostAPI().find(hostId, user, false);
-	}
-
-	private List<Contentlet> getContentletsByLink(final String linkId)
-			throws DotDataException, DotSecurityException {
-
-		final Link link = APILocator.getMenuLinkAPI()
-				.findWorkingLinkById(linkId, user, false);
-
-		if (link != null) {
-
-			if (link.getLinkType().equals(Link.LinkType.INTERNAL.toString())) {
-				final Identifier id = APILocator.getIdentifierAPI()
-						.find(link.getInternalLinkIdentifier());
-
-				// add file/content dependencies. will also work with htmlpages as content
-				if (InodeUtils.isSet(id.getInode()) && id.getAssetType()
-						.equals("contentlet")) {
-					return APILocator.getContentletAPI()
-									.search("+identifier:" + id.getId(), 0, 0, "moddate", user,
-											false);
-				}
-			}
-		}
-
-		return null;
-	}
-
-	private List<Contentlet> getContentletByLuceneQuery(final String luceneQuery)
-			throws DotDataException, DotSecurityException {
-		return APILocator.getContentletAPI()
-				.search(luceneQuery, 0, 0, null, user, false);
-	}
-
-	private List<Template> getTemplatesByHost(final Host host)
-			throws DotDataException, DotSecurityException {
-		return APILocator.getTemplateAPI().findTemplatesAssignedTo(host);
-	}
-
-	private List<Rule> getRulesByHost(final Host host)
-			throws DotDataException, DotSecurityException {
-		return APILocator.getRulesAPI().getAllRulesByParent(host, user, false);
-	}
-
-	private List<Folder> getFoldersByHost(final Host host)
-			throws DotDataException, DotSecurityException {
-		return APILocator.getFolderAPI()
-				.findFoldersByHost(host, user, false);
-	}
-
-	private List<Structure> getContentTypeByHost(final Host host)
-			throws DotDataException, DotSecurityException {
-		return StructureFactory
-				.getStructuresUnderHost(host, user, false);
-	}
-
-	private List<Structure> getContentTypeByFolder(final Folder folder)
-			throws DotDataException, DotSecurityException {
-		return APILocator.getFolderAPI().getStructures(folder, user, false);
-	}
-
-	private List<Container> getContainersByHost(final Host host) throws DotDataException {
-		return APILocator.getContainerAPI().findContainersUnder(host);
-	}
-
-	private List<FileAssetContainer> getFileContainersByHost(final Host host)
-			throws DotDataException {
-		return getContainersByHost(host).stream()
-				.filter(FileAssetContainer.class::isInstance)
-				.map(FileAssetContainer.class::cast)
-				.collect(Collectors.toList());
-	}
-
-	private Folder getParentFolder(Folder folder) throws DotDataException, DotSecurityException {
-		return APILocator.getFolderAPI().findParentFolder(folder, user, false);
-	}
-
-	private List<Link> getLinksByFolder(Folder folder)
-			throws DotDataException, DotSecurityException {
-		return APILocator.getMenuLinkAPI().findFolderMenuLinks(folder);
-	}
-
-	private Identifier findIdentifier(final String id) {
-		try {
-			return APILocator.getIdentifierAPI().find(id);
-		} catch (DotDataException e) {
-			Logger.error(this, e.getMessage(), e);
-			return null;
-		}
-	}
-
-	private List<Contentlet> findContentletsByIdentifier(final Identifier identifier) {
-		try {
-			return APILocator.getContentletAPI()
-					.findAllVersions(identifier, false, user, false);
-		} catch (DotDataException | DotSecurityException e) {
-			Logger.error(this, e.getMessage(), e);
-			return Collections.emptyList();
-		}
-	}
-
-	private List<Contentlet> getContentletsByPage(final IHTMLPage page)
-			throws DotDataException, DotSecurityException {
-
-		final Table<String, String, Set<PersonalizedContentlet>> pageMultiTrees =
-				APILocator.getMultiTreeAPI().getPageMultiTrees(page, false);
-
-		return pageMultiTrees.values().stream()
-				.flatMap(personalizedContentlets -> personalizedContentlets.stream())
-				.map(personalizedContentlet -> personalizedContentlet.getContentletId())
-				.filter(Objects::isNull)
-				.map(contentletId -> findIdentifier(contentletId))
-				.filter(Objects::isNull)
-				.flatMap(identifier -> findContentletsByIdentifier(identifier).stream())
-				.collect(Collectors.toList());
-	}
-
-	private List<Rule> getRuleByPage(final IHTMLPage page)
-			throws DotDataException, DotSecurityException {
-		return APILocator.getRulesAPI().getAllRulesByParent(page, user, false);
-	}
-
-	private Host getHostByTemplate(final Template template)
-			throws DotDataException, DotSecurityException {
-		return APILocator.getHostAPI()
-				.find(APILocator.getTemplateAPI().getTemplateHost(template)
-						.getIdentifier(), user, false);
-	}
-
-	private List<Container> getContainerByTemplate(final Template template)
-			throws DotDataException, DotSecurityException {
-		return APILocator.getTemplateAPI()
-				.getContainersInTemplate(template, user, false);
-	}
-
-	private Folder getThemeByTemplate(final Template template)
-			throws DotDataException, DotSecurityException {
-
-			final Folder themeFolder = APILocator.getFolderAPI()
-					.find(template.getTheme(), user, false);
-
-			return themeFolder;
-	}
-
-	private Host getHostByContainer(final Container container)
-			throws DotDataException, DotSecurityException {
-		return APILocator.getContainerAPI().getParentHost(container, user, false);
-	}
-
-	private Collection<Structure> getContentTypeByLiveContainer(final String containerId)
-			throws DotDataException, DotSecurityException {
-		return getContentTypeByContainer(containerId, true);
-	}
-
-	private Collection<Structure> getContentTypeByWorkingContainer(final String containerId)
-			throws DotDataException, DotSecurityException {
-		return getContentTypeByContainer(containerId, false);
-	}
-
-	private Collection<Structure> getContentTypeByContainer(final String containerId, final boolean live)
-			throws DotDataException, DotSecurityException {
-
-		final Container container = live ? APILocator.getContainerAPI()
-				.getLiveContainerById(containerId, user, false)
-				: APILocator.getContainerAPI()
-						.getWorkingContainerById(containerId, user, false);
-
-		if (container == null) {
-			return Collections.emptyList();
-		}
-
-		return APILocator.getContainerAPI().getContainerStructures(container).stream()
-				.map(containerStructure ->
-						CacheLocator.getContentTypeCache()
-								.getStructureByInode(containerStructure.getStructureId()))
-				.collect(Collectors.toSet());
-	}
-
-	private Folder getFolderById(final String id) throws DotDataException, DotSecurityException {
-		return APILocator.getFolderAPI().find(id, user, false);
-	}
-
-	/**
-	 * Given that a FileAssetTemplate is defined by a bunch of files we need to collect the folder that enclose'em
-	 */
-	private Folder getFileAssetTemplateRootFolder(final FileAssetTemplate fileAssetTemplate)
-			throws DotDataException, DotSecurityException {
-		try {
-			final String path = fileAssetTemplate.getPath();
-			return APILocator.getFolderAPI()
-						.findFolderByPath(path, fileAssetTemplate.getHost(), user, false);
-		}catch (DotSecurityException | DotDataException e) {
-			Logger.error(DependencyManager.class, "Error Collecting the Folder of the File Asset Template: " + fileAssetTemplate.getIdentifier(), e);
-			return null;
-		}
-	}
-
-	private Collection<WorkflowScheme> getWorkflowSchemasByContentType(final Structure structure){
-		try{
-			return APILocator.getWorkflowAPI().findSchemesForStruct(structure);
-		} catch (DotDataException e) {
-			Logger.debug(getClass(),
-					() -> "Could not get the Workflow Scheme Dependency for Structure ID: "
-							+ structure.getInode());
-			return null;
-		}
-	}
-
-	public <T> boolean isExcludeByFilter(final PusheableAsset pusheableAsset) {
-
-		return (PusheableAsset.RELATIONSHIP == pusheableAsset && !publisherFilter.isRelationships()) ||
-				publisherFilter.doesExcludeDependencyClassesContainsType(pusheableAsset.getType());
-	}
-
-	public <T> void tryToAddAllAndProcessDependencies(
-			final PusheableAsset pusheableAsset, final SupplierWithException<Collection<T>> getter)
-			throws DotDataException, DotSecurityException {
-
-		if (!isExcludeByFilter(pusheableAsset)) {
-			getter.get().stream().forEach(asset -> {
-					try {
-						tryToAdd(pusheableAsset, asset);
-						config.addWithDependencies(asset, pusheableAsset);
-					} catch (AssetExcludeByFilterException e) {
-						//ignore
-					} catch (AssetExcludeException e) {
-						dependencyProcessor.addAsset(asset, pusheableAsset);
-					}
-				});
-		}
-	}
-
-	public <T> Collection<T> tryToAddAll(
-			final PusheableAsset pusheableAsset, final SupplierWithException<Collection<T>> getter)
-			throws DotDataException, DotSecurityException {
-
-		if (!isExcludeByFilter(pusheableAsset)) {
-			final Collection<T> assets = getter.get();
-
-			if (assets != null) {
-				return assets.stream()
-					.filter(asset -> {
-						try {
-							return tryToAdd(pusheableAsset, asset);
-						} catch (AssetExcludeException e) {
-							return false;
-						}
-					})
-					.collect(Collectors.toSet());
-			}
-		}
-
-		return Collections.emptySet();
-	}
-
-	public <T> void tryToAddAndProcessDependencies(
-			final PusheableAsset pusheableAsset, final SupplierWithException<T> getter)
-			throws DotDataException, DotSecurityException {
-		try {
-			tryToAdd(pusheableAsset, getter).ifPresent(asset ->
-					config.addWithDependencies(asset, pusheableAsset));
-		} catch (AssetExcludeByFilterException e) {
-			//ignore
-		} catch (AssetExcludeException e) {
-			final T asset = getter.get();
-			dependencyProcessor.addAsset(asset, pusheableAsset);
-		}
-	}
-
-	public <T> Optional<T> tryToAddSilently (
-			final PusheableAsset pusheableAsset, final SupplierWithException<T> getter)
-			throws DotDataException, DotSecurityException{
-		try {
-			return tryToAdd(pusheableAsset, getter);
-		} catch (AssetExcludeException e) {
-			return Optional.empty();
-		}
-	}
-
-	public <T> Optional<T> tryToAdd(
-			final PusheableAsset pusheableAsset, final SupplierWithException<T> getter)
-			throws DotDataException, DotSecurityException, AssetExcludeException{
-
-		if (!isExcludeByFilter(pusheableAsset)) {
-			final T asset = getter.get();
-
-			if (asset == null) {
-				return Optional.empty();
-			}
-
-			return tryToAdd(pusheableAsset, asset) ? Optional.of(asset) : Optional.empty();
-		} else {
-			throw new AssetExcludeByFilterException(String.format("Exclude by Operation %s",
-					config.getOperation()));
-		}
-	}
-
-	private synchronized <T> boolean tryToAdd(final PusheableAsset pusheableAsset, final T asset)
-		throws AssetExcludeException {
-
-		if ( config.getOperation() != Operation.PUBLISH ) {
-			this.pushedAssetUtil.removePushedAssetForAllEnv(asset, pusheableAsset);
-			throw new AssetExcludeException(String.format("Exclude by Operation %s",
-					config.getOperation()));
-		}
-
-		if (Contentlet.class.isInstance(asset) && !Contentlet.class.cast(asset).isHost() &&
-				publisherFilter.doesExcludeDependencyQueryContainsContentletId(
-						((Contentlet) asset).getIdentifier())) {
-
-			throw new AssetExcludeByFilterException(String.format("Exclude by Contentlet Id Filter: %s",
-					((Contentlet) asset).getIdentifier()));
-		}
-
-		if (!shouldCheckModDate(asset) ||
-				!dependencyModDateUtil.excludeByModDate(asset, pusheableAsset)) {
-
-			return add(pusheableAsset, asset);
-		} else {
-			throw new AssetExcludeException(String.format("Exclude by Moddate"));
-		}
-	}
-
-	private <T> boolean add(final PusheableAsset pusheableAsset, final T asset) {
-		final boolean isAdded = config.add(asset, pusheableAsset);
-
-		if (isAdded) {
+		if (added) {
 			pushedAssetUtil.savePushedAssetForAllEnv(asset, pusheableAsset);
-		}
-
-		return isAdded;
-	}
-
-	private <T> boolean shouldCheckModDate(T asset) {
-		return !Language.class.isInstance(asset);
-	}
-
-	/**
-	 * For given Links adds its dependencies:
-	 * <ul>
-	 * <li>Hosts</li>
-	 * <li>Folders</li>
-	 * </ul>
-	 */
-	private void processLinkDependency(final Link link)  {
-		final String linkId = link.getIdentifier();
-
-		try {
-			Identifier ident = APILocator.getIdentifierAPI().find(linkId);
-
-			// Folder Dependencies
-			tryToAddSilently(PusheableAsset.FOLDER, () -> getFolderByParentIdentifier(ident));
-
-			// Host Dependencies
-			tryToAddSilently(PusheableAsset.SITE, () -> getHostById(ident.getHostId()));
-
-			// Content Dependencies
-			tryToAddAllAndProcessDependencies(PusheableAsset.CONTENTLET, () -> getContentletsByLink(linkId));
-		} catch (Exception e) {
-			Logger.error(this, "can't load menuLink deps "+linkId,e);
-		}
-	}
-
-	/**
-	 * Collects the different dependent objects that are required for pushing
-	 * {@link Host} objects. The required dependencies of a site are:
-	 * <ul>
-	 * <li>Templates.</li>
-	 * <li>Containers.</li>
-	 * <li>Contentlets.</li>
-	 * <li>Content Types.</li>
-	 * <li>Folders.</li>
-	 * <li>Rules.</li>
-	 * </ul>
-	 */
-	private void proccessHostDependency (final Host host) {
-		try {
-			// Template dependencies
-			tryToAddAllAndProcessDependencies(PusheableAsset.TEMPLATE,
-					() -> getTemplatesByHost(host));
-
-			// Container dependencies
-			tryToAddAllAndProcessDependencies(PusheableAsset.CONTAINER, () -> getContainersByHost(host));
-			getFileContainersByHost(host).stream()
-					.forEach(fileContainer -> dependencyProcessor.addAsset(fileContainer.getIdentifier(),
-							PusheableAsset.CONTAINER));
-
-			// Content dependencies
-			tryToAddAllAndProcessDependencies(PusheableAsset.CONTENTLET,
-					() -> getContentletByLuceneQuery("+conHost:" + host.getIdentifier()));
-
-			// Structure dependencies
-			tryToAddAllAndProcessDependencies(PusheableAsset.CONTENT_TYPE,
-					() -> getContentTypeByHost(host));
-
-			// Folder dependencies
-			tryToAddAllAndProcessDependencies(PusheableAsset.FOLDER,
-					() -> getFoldersByHost(host));
-
-			// Rule dependencies
-			tryToAddAllAndProcessDependencies(PusheableAsset.RULE,
-					() -> getRulesByHost(host));
-
-		} catch (DotSecurityException | DotDataException e) {
-			Logger.error(this, e.getMessage(),e);
-		}
-	}
-
-	/**
-	 * For given Folders adds its dependencies:
-	 * <ul>
-	 * <li>Hosts</li>
-	 * <li>Contentlets</li>
-	 * <li>Links</li>
-	 * <li>Structures</li>
-	 * <li>HTMLPages</li>
-	 * </ul>
-	 */
-	private void processFolderDependency(final Folder folder) {
-		try {
-			tryToAddSilently(PusheableAsset.FOLDER, () -> getParentFolder(folder));
-			setFolderListDependencies(folder);
-		} catch (DotSecurityException | DotDataException e) {
-			Logger.error(this, e.getMessage(),e);
-		}
-	}
-
-	/**
-	 *
-
-	 * @throws DotIdentifierStateException
-	 * @throws DotDataException
-	 * @throws DotSecurityException
-	 */
-	private void setFolderListDependencies(final Folder folder)
-			throws DotIdentifierStateException, DotDataException, DotSecurityException {
-
-		tryToAddSilently(PusheableAsset.SITE, () -> getHostById(folder.getHostId()));
-
-		// Content dependencies
-		tryToAddAllAndProcessDependencies(PusheableAsset.CONTENTLET,
-				() -> getContentletByLuceneQuery("+conFolder:" + folder.getInode()));
-
-		// Menu Link dependencies
-		tryToAddAllAndProcessDependencies(PusheableAsset.LINK,
-				() -> getLinksByFolder(folder));
-
-		// Structure dependencies
-		tryToAddAllAndProcessDependencies(PusheableAsset.CONTENT_TYPE,
-				() -> getContentTypeByFolder(folder));
-
-		//Add the default structure of this folder
-		tryToAddAndProcessDependencies(PusheableAsset.CONTENT_TYPE,
-				() -> CacheLocator.getContentTypeCache()
-						.getStructureByInode(folder.getDefaultFileType()));
-
-		// SubFolders
-		tryToAddAllAndProcessDependencies(PusheableAsset.FOLDER,
-				() -> APILocator.getFolderAPI().findSubFolders(folder, user, false));
-	}
-
-	/**
-	 * Collects the different dependent objects that are required for pushing
-	 * {@link IHTMLPage} objects. The required dependencies of a page are:
-	 * <ul>
-	 * <li>Host.</li>
-	 * <li>Template.</li>
-	 * <li>Containers.</li>
-	 * <li>Content Types.</li>
-	 * <li>Contentlets.</li>
-	 * <li>Rules.</li>
-	 * </ul>
-	 *
-	 */
-	private void processHTMLPagesDependency(final String pageId) {
-		try {
-
-			final IdentifierAPI idenAPI = APILocator.getIdentifierAPI();
-			final FolderAPI folderAPI = APILocator.getFolderAPI();
-			final Set<Container> containerList = new HashSet<>();
-
-			final Identifier identifier = idenAPI.find(pageId);
-
-			if(identifier==null || UtilMethods.isEmpty(identifier.getId())) {
-				Logger.warn(this.getClass(), "Unable to find page for identifier, moving on.  Id: " + identifier );
-				return;
-			}
-
-			// Host dependency
-			tryToAddSilently(PusheableAsset.SITE, () -> getHostById(identifier.getHostId()));
-
-			// Folder dependencies
-			tryToAddSilently(PusheableAsset.FOLDER, () -> getFolderByParentIdentifier(identifier));
-
-			// looking for working version (must exists)
-			final IHTMLPage workingPage = Try.of(
-					()->APILocator.getHTMLPageAssetAPI().findByIdLanguageFallback(
-							identifier,
-							APILocator.getLanguageAPI().getDefaultLanguage().getId(),
-							false,
-							user,
-							false)
-			).onFailure(e->Logger.warnAndDebug(DependencyManager.class, e)).getOrNull();
-
-			if(workingPage==null) {
-				return;
-			}
-
-			final IHTMLPage livePage = workingPage.isLive()
-					? workingPage
-					: Try.of(()->
-					APILocator.getHTMLPageAssetAPI().findByIdLanguageFallback(identifier, APILocator.getLanguageAPI().getDefaultLanguage().getId(), true, user, false))
-					.onFailure(e->Logger.warnAndDebug(DependencyManager.class, e)).getOrNull();
-
-			// working template working page
-			final Template workingTemplateWP = workingPage != null ?
-					APILocator.getTemplateAPI()
-							.findWorkingTemplate(workingPage.getTemplateId(), user, false) : null;
-
-			if (workingTemplateWP != null) {
-
-				// Templates dependencies
-				if(!(workingTemplateWP instanceof FileAssetTemplate)) {
-					tryToAddAndProcessDependencies(PusheableAsset.TEMPLATE, () -> workingTemplateWP);
-				} else {
-					dependencyProcessor.addAsset(workingTemplateWP, PusheableAsset.TEMPLATE);
-				}
-			}
-
-			final Template liveTemplateLP = livePage != null ?
-					APILocator.getTemplateAPI()
-							.findLiveTemplate(livePage.getTemplateId(), user, false) : null;
-
-			// Templates dependencies
-			if (liveTemplateLP != null ) {
-				if(!(liveTemplateLP instanceof FileAssetTemplate)) {
-					tryToAddAndProcessDependencies(PusheableAsset.TEMPLATE, () -> liveTemplateLP);
-				} else {
-					dependencyProcessor.addAsset(liveTemplateLP, PusheableAsset.TEMPLATE);
-				}
-			}
-
-			// Contents dependencies
-			tryToAddAllAndProcessDependencies(PusheableAsset.CONTENTLET, () -> getContentletsByPage(workingPage));
-
-
-			// Rule dependencies
-			tryToAddAllAndProcessDependencies(PusheableAsset.RULE, () -> getRuleByPage(workingPage));
-		} catch (DotSecurityException | DotDataException e) {
-			Logger.error(this, e.getMessage(),e);
-		}
-	}
-
-	private void processTemplateDependencies(final Template template) {
-
-		try {
-
-			final Template workingTemplate = template.isWorking() ? template
-			: APILocator.getTemplateAPI().findWorkingTemplate(template.getIdentifier(), user, false);
-			final Template liveTemplate = template.isLive() ? template :
-					APILocator.getTemplateAPI().findLiveTemplate(template.getIdentifier(), user, false);
-
-			// Host dependency
-			tryToAddSilently(PusheableAsset.SITE, () -> getHostByTemplate(workingTemplate));
-
-			addContainerByTemplate(workingTemplate);
-			addContainerByTemplate(liveTemplate);
-
-			tryToAddAndProcessDependencies(PusheableAsset.FOLDER, () -> getThemeByTemplate(workingTemplate));
-
-			if(workingTemplate instanceof FileAssetTemplate){
-				//Process FileAssetTemplate
-				tryToAddAndProcessDependencies(PusheableAsset.FOLDER, () ->
-						getFileAssetTemplateRootFolder(FileAssetTemplate.class.cast(workingTemplate)));
-			}
-		} catch (DotSecurityException | DotDataException e) {
-
-			Logger.error(this, e.getMessage(),e);
-		}
-	}
-
-	private void addContainerByTemplate(Template workingTemplate) throws DotDataException, DotSecurityException {
-		final List<Container> containerByTemplate = getContainerByTemplate(workingTemplate);
-
-		tryToAddAllAndProcessDependencies(PusheableAsset.CONTAINER, () ->
-					containerByTemplate.stream()
-						.filter(container -> !FileAssetContainer.class.isInstance(container))
-						.collect(Collectors.toList())
-		);
-
-		containerByTemplate.stream()
-			.filter(FileAssetContainer.class::isInstance)
-			.forEach(container -> dependencyProcessor.addAsset(container,
-					PusheableAsset.CONTAINER));
-	}
-
-
-
-	/**
-	 * For given Containers adds its dependencies:
-	 * <ul>
-	 * <li>Hosts</li>
-	 * <li>Structures</li>
-	 * </ul>
-	 */
-	private void processContainerDependency(final Container container)  {
-
-		try {
-			final String containerId = container.getIdentifier();
-			final Container containerById = APILocator.getContainerAPI()
-					.getWorkingContainerById(containerId, user, false);
-
-			// Host Dependency
-			tryToAddSilently(PusheableAsset.SITE, () -> getHostByContainer(containerById));
-
-			// Content Type Dependencies
-			tryToAddAllAndProcessDependencies(PusheableAsset.CONTENT_TYPE,
-					() -> getContentTypeByWorkingContainer(containerId));
-
-			tryToAddAllAndProcessDependencies(PusheableAsset.CONTENT_TYPE,
-					() -> getContentTypeByLiveContainer(containerId));
-
-			if (containerById instanceof FileAssetContainer ) {
-				// Process FileAssetContainer
-				tryToAddAndProcessDependencies(PusheableAsset.FOLDER, () ->
-						getFileAssetContainerRootFolder(FileAssetContainer.class.cast(containerById)));
-			}
-
-		} catch (DotSecurityException e) {
-
-			Logger.error(this, e.getMessage(),e);
-		} catch (DotDataException e) {
-			Logger.error(this, e.getMessage(),e);
-		}
-
-	}
-
-	/**
-	 * Function that wraps calling getWorkingContainerById
-	 */
-	private Function<String, Container> workingContainerById = id -> {
-		try {
-			return APILocator.getContainerAPI().getWorkingContainerById(id, user, false);
-		} catch (DotDataException | DotSecurityException e) {
-			Logger.error(this, e.getMessage(),e);
-		}
-		return null;
-	};
-
-	/**
-	 * Aux Predicate to simplify filtering FileAssetContainer
-	 */
-	private Predicate<Container> fileAssetContainer = container -> container instanceof FileAssetContainer;
-
-	/**
-	 * Given that a FileAssetContainer is defined by a bunch of vtl files we need to collect the folder that enclose'em
-	 * @param fileAssetContainer
-	 * @return
-	 */
-	private Folder getFileAssetContainerRootFolder(final FileAssetContainer fileAssetContainer) {
-		try {
-			final String path = fileAssetContainer.getPath();
-			return APILocator.getFolderAPI()
-						.findFolderByPath(path, fileAssetContainer.getHost(), user, false);
-		}catch (DotSecurityException | DotDataException e) {
-			Logger.error(DependencyManager.class, e);
-			return null;
-		}
-	}
-
-
-	public void processRelationshipDependencies(final Relationship relationship) {
-
-		try {
-			tryToAddAndProcessDependencies(PusheableAsset.CONTENT_TYPE,
-					() -> CacheLocator.getContentTypeCache()
-							.getStructureByInode(relationship.getChildStructureInode()));
-
-			tryToAddAndProcessDependencies(PusheableAsset.CONTENT_TYPE,
-					() -> CacheLocator.getContentTypeCache()
-							.getStructureByInode(relationship.getParentStructureInode()));
-		} catch (DotDataException | DotSecurityException e) {
-			Logger.error(this, e.getMessage(),e);
-		}
-	}
-
-	/**
-	 *
-	 * @throws DotDataException
-	 * @throws DotSecurityException
-	 */
-	private void processContentTypeDependency(final Structure structure) {
-		try{
-			// Host Dependency
-			tryToAddSilently(PusheableAsset.SITE, () -> getHostById(structure.getHost()));
-
-			// Folder Dependencies
-			tryToAddSilently(PusheableAsset.FOLDER, () -> getFolderById(structure.getFolder()));
-
-			// Workflows Dependencies
-			tryToAddAll(PusheableAsset.WORKFLOW, () -> getWorkflowSchemasByContentType(structure));
-
-			// Categories Dependencies
-			tryToAddAll(PusheableAsset.CATEGORY, () -> APILocator.getCategoryAPI()
-							.findCategories(new StructureTransformer(structure).from(), user));
-
-			// Related structures
-			tryToAddAllAndProcessDependencies(PusheableAsset.RELATIONSHIP, () ->
-					APILocator.getRelationshipAPI().byContentType(structure));
-
-		} catch (DotDataException | DotSecurityException e) {
-			Logger.error(this, e.getMessage(),e);
-		}
-	}
-
-	/**
-	 *
-	 * @throws DotDataException
-	 * @throws DotSecurityException
-	 */
-	private void processContentDependency(final Contentlet contentlet)
-			throws  DotBundleException {
-
-		try {
-			final String contentId = contentlet.getIdentifier();
-			final Identifier ident = APILocator.getIdentifierAPI().find(contentId);
-			final List<Contentlet> contentList =
-					APILocator.getContentletAPI().findAllVersions(ident, false, user, false);
-
-			final Set<Contentlet> contentsToProcess = new HashSet<Contentlet>();
-			final Set<Contentlet> contentsWithDependenciesToProcess = new HashSet<Contentlet>();
-
-			for (final Contentlet contentletVersion : contentList) {
-
-				if (contentletVersion.isHTMLPage()) {
-					processHTMLPagesDependency(contentletVersion.getIdentifier());
-				}
-
-				// Host Dependency
-				tryToAddSilently(PusheableAsset.SITE, () -> getHostById(contentletVersion.getHost()));
-
-				contentsToProcess.add(contentletVersion);
-
-				// Relationships Dependencies
-				final Map<Relationship, List<Contentlet>> contentRelationships = APILocator
-						.getContentletAPI().findContentRelationships(contentletVersion, user);
-
-				tryToAddAllAndProcessDependencies(PusheableAsset.RELATIONSHIP, () ->
-						contentRelationships.keySet());
-
-				if(publisherFilter.isRelationships()) {
-					contentRelationships.values().stream()
-							.forEach(contentlets -> contentsToProcess.addAll(contentlets));
-				}
-
-			}
-
-			for (final Contentlet contentletToProcess : contentsToProcess) {
-				// Host Dependency
-				tryToAddSilently(PusheableAsset.SITE, () -> getHostById(contentletToProcess.getHost()));
-
-				contentsWithDependenciesToProcess.add(contentletToProcess);
-				//Copy asset files to bundle folder keeping original folders structure
-				final List<Field> fields=FieldsCache.getFieldsByStructureInode(contentletToProcess.getStructureInode());
-
-				for(final Field field : fields) {
-					if (field.getFieldType().equals(Field.FieldType.IMAGE.toString())
-							|| field.getFieldType().equals(Field.FieldType.FILE.toString())) {
-
-						try {
-							String value = "";
-							if(UtilMethods.isSet(APILocator.getContentletAPI().getFieldValue(contentletToProcess, field))){
-								value = APILocator.getContentletAPI().getFieldValue(contentletToProcess, field).toString();
-							}
-							final Identifier id = APILocator.getIdentifierAPI().find(value);
-							if (InodeUtils.isSet(id.getInode()) && id.getAssetType().equals("contentlet")) {
-								contentsWithDependenciesToProcess.addAll(APILocator.getContentletAPI().findAllVersions(id, false, user, false));
-							}
-						} catch (Exception ex) {
-							Logger.debug(this, ex.toString());
-							throw new DotStateException("Problem occured while publishing file:" +ex.getMessage(), ex );
-						}
-					}
-
-				}
-			}
-
-			// Adding the Contents (including related) and adding filesAsContent
-			for (final Contentlet contentletWithDependenciesToProcess : contentsWithDependenciesToProcess) {
-				// Host Dependency
-				tryToAddSilently(PusheableAsset.SITE, () -> getHostById(contentletWithDependenciesToProcess.getHost()));
-
-				// Content Dependency
-				tryToAddSilently(PusheableAsset.CONTENTLET, () -> contentletWithDependenciesToProcess);
-
-				// Folder Dependency
-				tryToAddSilently(PusheableAsset.FOLDER, () -> getFolderById(contentletWithDependenciesToProcess.getFolder()));
-
-				// Language Dependency
-				tryToAddAndProcessDependencies(PusheableAsset.LANGUAGE, () ->
-						APILocator.getLanguageAPI().getLanguage(contentletWithDependenciesToProcess.getLanguageId())
-				);
-
-				try {
-					if (Config.getBooleanProperty("PUSH_PUBLISHING_PUSH_ALL_FOLDER_PAGES", false)
-							&& contentletWithDependenciesToProcess.getStructure().getStructureType() == Structure.STRUCTURE_TYPE_HTMLPAGE) {
-
-						final Folder contFolder=APILocator.getFolderAPI()
-								.find(contentletWithDependenciesToProcess.getFolder(), user, false);
-
-						tryToAddAllAndProcessDependencies(PusheableAsset.CONTENTLET, () -> {
-							final List<IHTMLPage> liveHtmlPages = APILocator.getHTMLPageAssetAPI()
-									.getHTMLPages(contFolder, false, false, user, false);
-
-							final List<IHTMLPage> workingHtmlPages = APILocator.getHTMLPageAssetAPI()
-									.getHTMLPages(contFolder, true, false, user, false);
-
-							return Stream.concat(liveHtmlPages.stream(), workingHtmlPages.stream())
-										.collect(Collectors.toSet());
-						});
-					}
-				} catch (Exception e) {
-					Logger.debug(this, e.toString());
-				}
-
-				if(Config.getBooleanProperty("PUSH_PUBLISHING_PUSH_STRUCTURES", true)) {
-					tryToAddAndProcessDependencies(PusheableAsset.CONTENT_TYPE, () ->
-								CacheLocator.getContentTypeCache()
-										.getStructureByInode(contentletWithDependenciesToProcess.getStructureInode())
-							);
-
-				}
-
-				// Evaluate all the categories from this  to include as dependency.
-				tryToAddAll(PusheableAsset.CATEGORY, () ->
-						APILocator.getCategoryAPI()
-								.getParents(contentletWithDependenciesToProcess, APILocator.systemUser(), false)
-				);
-
-			}
-		} catch (Exception e) {
-			throw new DotBundleException(this.getClass().getName() + " : " + "generate()"
-					+ e.getMessage() + ": Unable to pull content", e);
-		}
-	}
-
-	private void processLanguage(final Language language) throws DotBundleException {
-		try{
-			final long lang = language.getId();
-			final String keyValueQuery = "+contentType:" + LanguageVariableAPI.LANGUAGEVARIABLE + " +languageId:" + lang;
-			final List<Contentlet> listKeyValueLang = APILocator.getContentletAPI()
-					.search(keyValueQuery,0, -1, StringPool.BLANK, user, false);
-			tryToAddAll(PusheableAsset.CONTENTLET, () -> listKeyValueLang);
-
-			final String contentTypeId = listKeyValueLang.get(0).getContentTypeId();
-			tryToAddSilently(PusheableAsset.CONTENT_TYPE, () ->
-					CacheLocator.getContentTypeCache().getStructureByInode(contentTypeId));
-
-		} catch (Exception e) {
-			throw new DotBundleException(this.getClass().getName() + " : " + "generate()"
-					+ e.getMessage() + ": Unable to pull content", e);
-		}
-	}
-
-	/**
-	 * Collects the different dependent objects that are required for pushing
-	 * {@link Rule} objects. The required dependency of a rule is either:
-	 * <ol>
-	 * <li>The Site (host) they were created in.</li>
-	 * <li>Or the Content Page they were created in.</li>
-	 * </ol>
-	 */
-	private void setRuleDependencies(final Rule rule)  {
-		String ruleToProcess = "";
-		final HostAPI hostAPI = APILocator.getHostAPI();
-		final ContentletAPI contentletAPI = APILocator.getContentletAPI();
-
-		try {
-			ruleToProcess = rule.getId();
-			final List<Contentlet> contentlets = contentletAPI.searchByIdentifier(
-					"+identifier:" + rule.getParent(), 1, 0, null, this.user, false,
-					PermissionAPI.PERMISSION_READ, true);
-
-			if (contentlets != null && contentlets.size() > 0) {
-				final Contentlet parent = contentlets.get(0);
-
-				if (parent.isHost()) {
-					tryToAddSilently(PusheableAsset.SITE,
-							() -> hostAPI.find(rule.getParent(), this.user, false));
-				} else if (parent.isHTMLPage()) {
-					tryToAddSilently(PusheableAsset.CONTENTLET, () -> parent);
-				} else {
-					throw new DotDataException("The parent ID [" + parent.getIdentifier() + "] is a non-valid parent.");
-				}
-			} else {
-				throw new DotDataException("The parent ID [" + rule.getParent() + "] cannot be found for Rule [" + rule.getId() + "]");
-			}
-		} catch (DotDataException e) {
-			Logger.error(this, "Dependencies for rule [" + ruleToProcess + "] could not be set: " + e.getMessage(), e);
-		} catch (DotSecurityException e) {
-			Logger.error(this, "Dependencies for rule [" + ruleToProcess + "] could not be set: " + e.getMessage(), e);
-		}
-	}
-
-	private void setLanguageVariables() {
-		final ContentletAPI contentletAPI = APILocator.getContentletAPI();
-		final Date date = new Date();
-		try{
-			//We're no longer filtering by language here..
-			//The reason is We're simply collecting all available lang variables so we can infer additional languages used. see #15359
-			final String langVarsQuery = "+contentType:" + LanguageVariableAPI.LANGUAGEVARIABLE ;
-			final List<Contentlet> langVariables = contentletAPI.search(langVarsQuery, 0, -1, StringPool.BLANK, user, false);
-
-			tryToAddAllAndProcessDependencies(PusheableAsset.CONTENTLET, () -> langVariables);
-		}catch (Exception e){
-			Logger.error(this, e.getMessage(),e);
 		}
 	}
